requirements.txt
mysql-connector-python==9.2.0
numpy==2.2.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==11.1.0
